+++ src/include/Flags.hpp
@@ -10,3 +10,5 @@ void registerFastFlags(std::unordered_map<std::string, std::string>& fastFlags, 
 
 /// Register FFlags but emit errors straight to stderr
 void registerFastFlagsCLI(std::unordered_map<std::string, std::string>& fastFlags);
+
+void applyRequiredFlags();