+++ tests/Autocomplete.test.cpp
@@ -31,7 +31,7 @@ std::optional<lsp::CompletionItem> getItem(const std::vector<lsp::CompletionItem
 lsp::CompletionItem requireItem(const std::vector<lsp::CompletionItem>& items, const std::string& label)
 {
     auto item = getItem(items, label);
-    REQUIRE_MESSAGE(item.has_value(), "no item found");
+    REQUIRE_MESSAGE(item.has_value(), "no item found ", label);
     return item.value();
 }
 
@@ -830,10 +830,11 @@ TEST_CASE_FIXTURE(Fixture, "auto_imports_handles_multi_line_existing_requires_wh
     CHECK_EQ(insertedLineNumber, 3);
 }
 
-static void checkFileCompletionExists(const std::vector<lsp::CompletionItem>& items, const std::string& label)
+static void checkFileCompletionExists(const std::vector<lsp::CompletionItem>& items, const std::string& label, const std::string& insertText)
 {
     auto item = requireItem(items, label);
     CHECK_EQ(item.kind, lsp::CompletionItemKind::File);
+    CHECK_EQ(item.insertText, insertText);
 }
 
 static void checkFolderCompletionExists(const std::vector<lsp::CompletionItem>& items, const std::string& label)
@@ -842,7 +843,7 @@ static void checkFolderCompletionExists(const std::vector<lsp::CompletionItem>& 
     CHECK_EQ(item.kind, lsp::CompletionItemKind::Folder);
 }
 
-TEST_CASE_FIXTURE(Fixture, "string_require_default_shows_files_in_current_directory")
+TEST_CASE_FIXTURE(Fixture, "string_require_default_shows_initial_values")
 {
     TempDir t("require_default");
     t.write_child("first.luau", "return {}");
@@ -861,11 +862,9 @@ TEST_CASE_FIXTURE(Fixture, "string_require_default_shows_files_in_current_direct
 
     auto result = workspace.completion(params);
 
-    REQUIRE_EQ(result.size(), 4);
+    REQUIRE_EQ(result.size(), 2);
     checkFolderCompletionExists(result, "..");
-    checkFileCompletionExists(result, "source.luau");
-    checkFileCompletionExists(result, "first.luau");
-    checkFileCompletionExists(result, "second.luau");
+    checkFolderCompletionExists(result, ".");
 }
 
 TEST_CASE_FIXTURE(Fixture, "string_require_shows_files_in_current_directory")
@@ -889,9 +888,9 @@ TEST_CASE_FIXTURE(Fixture, "string_require_shows_files_in_current_directory")
 
     REQUIRE_EQ(result.size(), 4);
     checkFolderCompletionExists(result, "..");
-    checkFileCompletionExists(result, "source.luau");
-    checkFileCompletionExists(result, "first.luau");
-    checkFileCompletionExists(result, "second.luau");
+    checkFileCompletionExists(result, "source.luau", "source");
+    checkFileCompletionExists(result, "first.luau", "first");
+    checkFileCompletionExists(result, "second.luau", "second");
 }
 
 TEST_CASE_FIXTURE(Fixture, "string_require_shows_files_in_parent_directory")
@@ -916,8 +915,8 @@ TEST_CASE_FIXTURE(Fixture, "string_require_shows_files_in_parent_directory")
     REQUIRE_EQ(result.size(), 4);
     checkFolderCompletionExists(result, "..");
     checkFolderCompletionExists(result, "nested");
-    checkFileCompletionExists(result, "parent_first.luau");
-    checkFileCompletionExists(result, "parent_second.luau");
+    checkFileCompletionExists(result, "parent_first.luau", "parent_first");
+    checkFileCompletionExists(result, "parent_second.luau", "parent_second");
 }
 
 TEST_CASE_FIXTURE(Fixture, "string_require_shows_file_in_nested_directory")
@@ -941,7 +940,7 @@ TEST_CASE_FIXTURE(Fixture, "string_require_shows_file_in_nested_directory")
 
     REQUIRE_EQ(result.size(), 2);
     checkFolderCompletionExists(result, "..");
-    checkFileCompletionExists(result, "child.luau");
+    checkFileCompletionExists(result, "child.luau", "child");
 }
 
 TEST_CASE_FIXTURE(Fixture, "string_require_shows_files_in_current_directory_after_entering_then_exiting_child_folder")
@@ -966,8 +965,8 @@ TEST_CASE_FIXTURE(Fixture, "string_require_shows_files_in_current_directory_afte
     REQUIRE_EQ(result.size(), 4);
     checkFolderCompletionExists(result, "..");
     checkFolderCompletionExists(result, "nested");
-    checkFileCompletionExists(result, "source.luau");
-    checkFileCompletionExists(result, "sibling.luau");
+    checkFileCompletionExists(result, "source.luau", "source");
+    checkFileCompletionExists(result, "sibling.luau", "sibling");
 }
 
 TEST_CASE_FIXTURE(Fixture, "string_require_contains_luaurc_aliases")
@@ -996,65 +995,9 @@ TEST_CASE_FIXTURE(Fixture, "string_require_contains_luaurc_aliases")
 
     CHECK_EQ(result.size(), 4);
     checkFolderCompletionExists(result, "..");
-    checkFileCompletionExists(result, "source.luau");
+    checkFolderCompletionExists(result, ".");
     requireItem(result, "@Roact");
     requireItem(result, "@Fusion");
-}
-
-TEST_CASE_FIXTURE(Fixture, "string_require_contains_file_aliases")
-{
-    TempDir t("file_aliases");
-    client->globalConfig.require.fileAliases = {
-        {"@test1", "file1.luau"},
-        {"@test2", "file2.luau"},
-    };
-
-    auto [source, marker] = sourceWithMarker(R"(
-        --!strict
-        local x = require("|")
-    )");
-
-    auto uri = newDocument(t.touch_child("source.luau"), source);
-
-    lsp::CompletionParams params;
-    params.textDocument = lsp::TextDocumentIdentifier{uri};
-    params.position = marker;
-
-    auto result = workspace.completion(params);
-
-    CHECK_EQ(result.size(), 4);
-    checkFolderCompletionExists(result, "..");
-    checkFileCompletionExists(result, "source.luau");
-    checkFileCompletionExists(result, "@test1");
-    checkFileCompletionExists(result, "@test2");
-}
-
-TEST_CASE_FIXTURE(Fixture, "string_require_contains_directory_aliases")
-{
-    TempDir t("directory_aliases");
-    client->globalConfig.require.directoryAliases = {
-        {"@dir1", "directory1"},
-        {"@dir2", "directory2"},
-    };
-
-    auto [source, marker] = sourceWithMarker(R"(
-        --!strict
-        local x = require("|")
-    )");
-
-    auto uri = newDocument(t.touch_child("source.luau"), source);
-
-    lsp::CompletionParams params;
-    params.textDocument = lsp::TextDocumentIdentifier{uri};
-    params.position = marker;
-
-    auto result = workspace.completion(params);
-
-    CHECK_EQ(result.size(), 4);
-    checkFolderCompletionExists(result, "..");
-    checkFileCompletionExists(result, "source.luau");
-    checkFolderCompletionExists(result, "@dir1");
-    checkFolderCompletionExists(result, "@dir2");
 }
 
 TEST_CASE_FIXTURE(Fixture, "string_require_doesnt_show_aliases_after_a_directory_separator_is_seen")
@@ -1093,35 +1036,6 @@ TEST_CASE_FIXTURE(Fixture, "string_require_doesnt_show_aliases_after_a_directory
     checkFolderCompletionExists(result, "..");
 }
 
-TEST_CASE_FIXTURE(Fixture, "string_require_shows_files_under_a_directory_alias")
-{
-    TempDir t("lune");
-    t.write_child("process.luau", "return {}");
-    t.write_child("net.luau", "return {}");
-
-    client->globalConfig.require.directoryAliases = {
-        {"@lune", t.path()},
-    };
-
-    auto [source, marker] = sourceWithMarker(R"(
-        --!strict
-        local x = require("@lune/|")
-    )");
-
-    auto uri = newDocument("foo.luau", source);
-
-    lsp::CompletionParams params;
-    params.textDocument = lsp::TextDocumentIdentifier{uri};
-    params.position = marker;
-
-    auto result = workspace.completion(params);
-
-    REQUIRE_EQ(result.size(), 3);
-    checkFolderCompletionExists(result, "..");
-    checkFileCompletionExists(result, "process.luau");
-    checkFileCompletionExists(result, "net.luau");
-}
-
 TEST_CASE_FIXTURE(Fixture, "string_require_shows_files_under_a_luaurc_directory_alias")
 {
     TempDir t("lune");
@@ -1152,8 +1066,8 @@ TEST_CASE_FIXTURE(Fixture, "string_require_shows_files_under_a_luaurc_directory_
 
     REQUIRE_EQ(result.size(), 3);
     checkFolderCompletionExists(result, "..");
-    checkFileCompletionExists(result, "process.luau");
-    checkFileCompletionExists(result, "net.luau");
+    checkFileCompletionExists(result, "process.luau", "process");
+    checkFileCompletionExists(result, "net.luau", "net");
 }
 
 TEST_CASE_FIXTURE(Fixture, "autocomplete_end_for_incomplete_function")